% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato_font.R
\name{import_lato}
\alias{import_lato}
\alias{is_lato_imported}
\title{Import Lato Font}
\usage{
import_lato()

is_lato_imported()
}
\description{
\code{import_lato} makes the included Lato font available in R. This process
only needs to be completed once.

\code{is_lato_imported} checks if Lato has been imported
}
